project/requirements.txt
celery==4.4.7
Django==3.2.4
flower==0.9.7
redis==3.5.3

pytest==6.2.4
pytest-django==4.4.0
django-datatables-view

urllib3
beautifulsoup4
PyYAML